+++ src/nidaqlib/channels/__init__.py
@@ -2,7 +2,11 @@
 
 from __future__ import annotations
 
-from nidaqlib.channels.analog_input import AnalogInputVoltage, ThermocoupleInput
+from nidaqlib.channels.analog_input import (
+    AnalogInputBase,
+    AnalogInputVoltage,
+    ThermocoupleInput,
+)
 from nidaqlib.channels.analog_output import AnalogOutputVoltage
 from nidaqlib.channels.base import ChannelSpec, register_channel_kind
 from nidaqlib.channels.counter_input import (
@@ -19,6 +23,7 @@ from nidaqlib.channels.digital_input import DigitalInput
 from nidaqlib.channels.digital_output import DigitalOutput
 
 __all__ = [
+    "AnalogInputBase",
     "AnalogInputVoltage",
     "AnalogOutputVoltage",
     "ChannelSpec",